CSVParserTest.java

import static org.junit.Assert.*;
import org.junit.Before;
import org.junit.Test;
import java.util.List;

public class CSVParserTest {

    private CSVParser parser;

    @Before
    public void setUp() {
        parser = new CSVParser();
        // Additional setup if required
    }

    @Test
    public void testParseValidCSV() {
        // Assuming a sample CSV file "test_data.csv" with valid credit card data
        List<CreditCard> cards = parser.parseInputFile("input_file.csv");
        assertNotNull(cards);
        assertFalse(cards.isEmpty());

        // Further assertions can be added to check for specific card types, counts, etc.
    }

    // Additional tests can be added for different scenarios like empty file, invalid data, etc.
}
